refactor(providers): give the client and the deadlines their own file
`provider.rs` is at the structure limit, and the streaming fold that lands
alongside this pushes it over. Two subjects come out rather than one, because
the module holds three: what a provider *is*, how one is *called*, and how the
pieces of an answer go back together.

`provider/http.rs` takes the second - the deadlines a call runs under and the
client that carries them, which is one builder and one precedence order shared
by every provider so none can quietly acquire its own. Re-exported flat, so
`provider::build_http_client` and every other existing path still resolves and
this stays a pure move.
